new Modal component -> RemonteControlError uses it
